feat: blocks crm for case study

+++ src/utils/date.ts
@@ -2,7 +2,7 @@ export const formatDate = (dateStr?: string) => {
   if (!dateStr) return "";
 
   return new Date(dateStr).toLocaleDateString("en-US", {
-    month: "short",
+    month: "long",
     day: "numeric",
     year: "numeric",
   });
